// web-client/src/types.d.ts
export interface SubscriptionPlanType {
  id: number;
  name: string;
  price: float;
  duration: number;
  features: string | null;
  audioQuality: number | null;
  allowsDownloads: boolean;
  adFree: boolean;
}